+++ CHANGELOG.md
@@ -10,6 +10,7 @@ All notable changes to this project will be documented in this file, in reverse 
 * [#9](https://github.com/sandrokeil/interop-config/issues/9): Introducing HasDefaultOptions Interface
 * [#13](https://github.com/sandrokeil/interop-config/issues/13): Support for recursive mandatory options check
 * `canRetrieveOptions()` to `ConfigurationTrait` to perform the options check without throwing an exception 
+* `optionsWithFallback()` to `ConfigurationTrait` which uses default options if config can not be retrieved
 * OptionNotFoundException and MandatoryOptionNotFoundException extends OutOfBoundsException instead of RuntimeException
 * Check if retrieved options are an array or an instance of ArrayAccess
 * Benchmark suite